obs-transitions: Add narrowing casts

+++ plugins/obs-transitions/transition-stinger.c
@@ -88,7 +88,7 @@ static void stinger_update(void *data, obs_data_t *settings)
 
 	s->track_matte_enabled =
 		obs_data_get_bool(settings, "track_matte_enabled");
-	s->matte_layout = obs_data_get_int(settings, "track_matte_layout");
+	s->matte_layout = (int)obs_data_get_int(settings, "track_matte_layout");
 	s->matte_width_factor =
 		(s->matte_layout == MATTE_LAYOUT_HORIZONTAL ? 2.0f : 1.0f);
 	s->matte_height_factor =
@@ -507,7 +507,7 @@ static bool transition_point_type_modified(obs_properties_t *ppts,
 static bool track_matte_layout_modified(obs_properties_t *ppts,
 					obs_property_t *p, obs_data_t *s)
 {
-	int matte_layout = obs_data_get_int(s, "track_matte_layout");
+	int matte_layout = (int)obs_data_get_int(s, "track_matte_layout");
 	obs_property_t *prop_matte_path =
 		obs_properties_get(ppts, "track_matte_path");
